Refactored AlphaBeta search filters to replace `long` with `int` for evaluation values, removed `AlphaBetaHelper` encoding/decoding dependencies, and updated related methods, tests, and transposition logic accordingly.

+++ search/src/main/java/net/chesstango/search/smart/alphabeta/AlphaBetaFunction.java
@@ -5,5 +5,5 @@ package net.chesstango.search.smart.alphabeta;
  */
 @FunctionalInterface
 public interface AlphaBetaFunction {
-    long search(final int currentPly, final int alpha, final int beta);
+    int search(final int currentPly, final int alpha, final int beta);
 }
